Added publish/subscribe cache setting to GravityNode API.
Added boolean flag to receive last cached data product value when subscribing to a data product, so subscriber can opt out of receiving a cached data product.

 Added boolean flag to cache last sent data product or not when registering data product, so a publisher can choose to not cache the last sent data product value. 

Added support for setting default behavior for the above through an .ini setting with the configuration keys CacheLastSentDataproduct and ReceiveLastSentDataproduct. This will only reflect in the default behavior for the above methods, and explicitely providing a boolean for caching settings in GravityNode::Subscribe or GravityNode::RegisterDataProduct will honor that value.